+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+## 1.3.0
+### Breaking changes
+* Setting `enabled` is a number (previously boolean) <BR>
+  (0: never, 1: always, 2: limited to workspaceFolder)
+* Setting `exclude` is an array (previously string) <BR>
+  array of folder or files to not save (glob)
+### What's new ?
+* Save single files, without workspacefolder [#8](https://github.com/zabel-xyz/local-history/issues/8)
+* Setting `absolute` to allow saving absolute path in `local-history.path` [#16](https://github.com/zabel-xyz/local-history/issues/16)
+
 ## 1.2.1
 * Fix issue with variables like %AppData% in local-history.path [#15](https://github.com/zabel-xyz/local-history/issues/15)
 * Replace jade by pug